requirements.txt
PyQt5==5.15.9
python-docx==1.1.2
Pillow==10.4.0
requests==2.31.0
PyMuPDF>=1.23.0 